mod: build

+++ lib/commonjs/extensions/geo-export/mesh-exporter.d.ts
@@ -70,7 +70,7 @@ export declare abstract class MeshExporter<D extends RenderObjectExportData> imp
         vertices: Float32Array;
         normals: Float32Array | undefined;
         indices: Uint32Array | undefined;
-        groups: Float32Array | Uint8Array;
+        groups: Uint8Array | Float32Array;
         vertexCount: number;
         drawCount: number;
     };
